requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-bootstrap-v5==1.0.11
django-import-export==2.8.0
django-qr-code==3.1.0
et-xmlfile==1.1.0
heroku==0.1.4
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.1
pycparser==2.21
pydantic==1.9.1
python-dateutil==1.5
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
segno==1.5.2
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
